% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{laplace}
\alias{laplace}
\alias{dlaplace}
\alias{plaplace}
\alias{qlaplace}
\alias{rlaplace}
\title{Laplace distribution}
\usage{
dlaplace(x, mu = 0, scale = 1, log = FALSE)

plaplace(q, mu = 0, scale = 1)

qlaplace(p, mu = 0, scale = 1)

rlaplace(n, mu = 0, scale = 1)
}
\arguments{
\item{x}{vector of values.}

\item{mu}{location parameter.}

\item{scale}{scale parameter.}

\item{log}{flag for log density.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
A vector of density, distribution function, quantile or random values.
}
\description{
Laplace distribution
}
